+++ src/components/mainDisplay/sidebar/event.js
