+++ src/index.js
@@ -4,8 +4,8 @@ import PropTypes from 'prop-types'; // ES6
 import { BrowserRouter, Route } from 'react-router-dom'
 
 //COMPONENTs
-import Navbar from './components/navbar';
-import MainDisplay from './components/mainDisplay'
+import Navbar from './components/header/navbar';
+import MainDisplay from './components/mainDisplay/mainDisplay'
 
 const App = () =>{
     return (
